safari-600.7-branch/LayoutTests/fast/events/touch/gesture/pad-gesture-cancel-expected.txt

Tests basic use of GestureFlingCancel

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL actualWheelEventsOccurred should be 1. Was 0.
PASS successfullyParsed is true

TEST COMPLETE
